requirements-testing.txt
mock
mysqlclient
pylibmc
pytest-cov
pytest-pep8
